Add support for dynamic resource, static gophermap
Dynamic resources are any executable non-directory file.
The output of executing the file is served as a text resource.
Static gophermaps are traditional files-as-gophermaps; to use,
you will need to set the selector-type attribute to "1".

This version of dynamic resources is pretty much terrible. There
is no timeout, so a resource that hangs will hang geomyid. There
seems to be no clean way to fix this with the current architecture.
Good thing we're not too attached to the current architecture.